WI 10: Add treecutting mechanics <br> Rudimentary treecutting mechanics added. <br> Now it's necessary to add stockpile, because treelog is destroyed when woodcutter brings it home

+++ Assets/_Scripts/Units/IUnit.cs
@@ -7,7 +7,7 @@ public interface IUnit
     public static UnitSpawn OnUnitSpawned;
     public static UnitRemove OnUnitRemoved;
     string Name { get; set; }
-    float Health { get; set; }
+    int Health { get; set; }
     UnitState CurrentUnitState { get; set; }
     UnitType Type { get; set; }
     bool TakesAccomodation { get; }
